Moved App files to app folder in component; added required field to username and password
